Add subclasses to observed_classes when observer is instantiated
rails-observers was again getting into an infinite loop when loading the
device classes after moving to rails 5.1.  I'm now questioning if
it was working at all on rails 5.0.

So stop looking for descendents on every single model load.
Find all the subclasses once and be done with it.

We really really should think about getting far away from this gem.